Move DLCWalletApi to core module
