+++ app/server-test/src/test/scala/org/bitcoins/wallet/MockWalletApi.scala
@@ -1,9 +1,9 @@
 package org.bitcoins.wallet
 
+import org.bitcoins.core.api.dlc.wallet.AnyDLCHDWalletApi
 import org.bitcoins.core.api.wallet.db.AccountDb
 import org.bitcoins.core.hd.AddressType
 import org.bitcoins.core.protocol.BitcoinAddress
-import org.bitcoins.dlc.wallet.AnyDLCHDWalletApi
 
 import scala.concurrent.Future
